requirements.txt
google-genai[aiohttp]
pydantic
Pillow
pyautogui
python-mss
python-evdev
dbus-fast
robotframework
python-dotenv
